kvm create vps

+++ kvm/php/control.php
@@ -1,7 +1,9 @@
 <?php
 // https://www.linux.org.ru/forum/admin/12097297 (start non-root)
 // https://poiuty.com/index.php?title=%D0%9A%D0%BE%D0%BC%D0%BF%D0%B8%D0%BB%D0%B8%D1%80%D1%83%D0%B5%D0%BC_libvirt-php (build libvirt-php)
-$uri = "qemu+unix:///system"; $credentials = [VIR_CRED_AUTHNAME => "", VIR_CRED_PASSPHRASE => ""]; $res = libvirt_connect($uri,false,$credentials);
+function kvm_exec($command, $id){
+	shell_exec("virsh -c qemu:///system $command $id");
+}
 
 function get_status($name){
 	global $url, $credentials, $res;
@@ -9,25 +11,26 @@ function get_status($name){
 	return libvirt_domain_get_info($id); // 1 => online, 5 => offline
 }
 
-function kvm_exec($command, $id){
-	shell_exec("virsh -c $uri $command $id");
+$credentials = [VIR_CRED_AUTHNAME => "", VIR_CRED_PASSPHRASE => ""];
+$res = libvirt_connect("qemu+unix:///system", false, $credentials);
+
+if(empty(intval($_GET['id'])) || empty($_GET['command']) || empty($_GET['key'])) die("error 1");
+if($_GET['key'] != 'xxx') die("error 2");
+
+$vm_id = "kvm".intval($_GET['id']);
+$status = get_status($vm_id);
+
+switch($_GET['command']){
+	case 'stopServer':
+		kvm_exec("destroy", $vm_id);
+	break;
+	case 'startServer':
+		kvm_exec("start", $vm_id);
+	break;
+	case 'restartServer':
+		kvm_exec("reboot", $vm_id);
+	break;
 }
 
-if(empty(intval($_GET['id'])) || empty($_GET['command'])) die("error 1");
-$vm_id = "kvm".intval($_GET['id']); $status = get_status($vm_id);
-
-var_dump($status);
-
-if($_GET['command'] == 'stop'){
-	kvm_exec("destroy", $vm_id);
-}
-
-if($_GET['command'] == 'start'){
-	kvm_exec("start", $vm_id);
-}
-
-if($_GET['command'] == 'reboot'){
-	kvm_exec("reboot", $vm_id);
-}
-
-echo $status['state'];
+if($status['state'] == 1) echo 'running';
+	else echo 'down';
